+++ src/conformist.ml
@@ -176,7 +176,7 @@ let rec fold_left'
 
 let fold_left ~f ~init schema = fold_left' ~f ~init schema.fields
 
-type validation_error = (string * string) list
+type error = string * string option * string
 type input = (string * string list) list
 
 let validate schema input =
@@ -185,17 +185,19 @@ let validate schema input =
     match List.assoc name input with
     | [ value_string ] ->
       (match Field.validate field value_string with
-      | Some msg -> List.cons (name, msg) errors
+      | Some msg -> List.cons (name, Some value_string, msg) errors
       | None -> errors)
-    | _ -> List.cons (name, "Multiple values provided") errors
+    | values ->
+      let value = Format.sprintf "[%s]" (String.concat ", " values) in
+      List.cons (name, Some value, "Multiple values provided") errors
     | exception Not_found ->
       (match Field.optional field, Field.encode_default field with
       | _, Some default ->
         (match Field.validate field default with
-        | Some msg -> List.cons (name, msg) errors
+        | Some msg -> List.cons (name, None, msg) errors
         | None -> errors)
       | true, None -> errors
-      | false, None -> List.cons (name, "No value provided") errors)
+      | false, None -> List.cons (name, None, "No value provided") errors)
   in
   fold_left ~f ~init:[] schema |> List.rev
 ;;
@@ -228,7 +230,7 @@ let rec decode
         | ctor -> decode { fields; ctor } fields_assoc
         | exception exn ->
           let msg = Printexc.to_string exn in
-          Error (field.name, Some "", msg))
+          Error (field.name, None, msg))
       | None ->
         (match field.decoder "" with
         | Ok value ->
@@ -236,9 +238,11 @@ let rec decode
           | ctor -> decode { fields; ctor } fields_assoc
           | exception exn ->
             let msg = Printexc.to_string exn in
-            Error (field.name, Some "", msg))
+            Error (field.name, None, msg))
         | Error msg -> Error (field.name, Some "", msg)))
-    | _ -> Error (field.name, None, "Multiple values provided")
+    | values ->
+      let value = Format.sprintf "[%s]" (String.concat ", " values) in
+      Error (field.name, Some value, "Multiple values provided")
     | exception Not_found ->
       (match field.default with
       | Some value ->
@@ -246,7 +250,8 @@ let rec decode
         | ctor -> decode { fields; ctor } fields_assoc
         | exception exn ->
           let msg = Printexc.to_string exn in
-          Error (field.name, None, msg))
+          let value_string = Option.map field.encoder field.default in
+          Error (field.name, value_string, msg))
       | None -> Error (field.name, None, "No value provided")))
 ;;
 
@@ -255,9 +260,9 @@ let decode_and_validate schema input =
   match decode schema input, validation_errors with
   | Ok value, [] -> Ok value
   | Ok _, validation_errors -> Error validation_errors
-  | Error (field_name, _, msg), validation_errors ->
+  | Error (field_name, value, msg), validation_errors ->
     validation_errors
-    |> List.filter (fun (name, _) -> not (String.equal name field_name))
-    |> List.cons (field_name, msg)
+    |> List.filter (fun (name, _, _) -> not (String.equal name field_name))
+    |> List.cons (field_name, value, msg)
     |> Result.error
 ;;