+++ src/conformist.mli
@@ -302,8 +302,15 @@ val fold_left
   -> ('meta, 'args, 'ty) t
   -> 'res
 
-(** An empty [validation_error] means that the schema is valid. *)
-type validation_error = (string * string) list
+(** An error [(field, value, error_message)] is used to for decoding errors and
+    validation errors.
+
+    [field] is the field name of the input that failed to decode or validate,
+    [value] is the input value (if one was provided) and [error_message] is the
+    decoding or validation error message.
+
+    An empty list of [error] means that the schema is valid. *)
+type error = string * string option * string
 
 (** The [input] represents unsafe data that needs to be decoded and validated.
     This is typically some user input. *)
@@ -312,16 +319,8 @@ type input = (string * string list) list
 (** [decode schema input] returns the decoded value of type ['ty] by decoding
     the [input] using the [schema].
 
-    The returned error value is a triple [(field_name, input_value, error_msg)].
-    [field_name] is the field name of the input that failed to decode,
-    [input_value] is the input value if one was provided and [error_msg] is the
-    error message.
-
     No validation logic is executed in this step. *)
-val decode
-  :  ('meta, 'ctor, 'ty) t
-  -> input
-  -> ('ty, string * string option * string) result
+val decode : ('meta, 'ctor, 'ty) t -> input -> ('ty, error) result
 
 (** [validate schema input] returns a list of validation errors by running the
     validators defined in [schema] on the [input] data. An empty list implies
@@ -330,15 +329,11 @@ val decode
 
     Note that [input] that has no validation errors might still fail to decode,
     depending on the validation functions specified in [schema]. *)
-val validate : ('meta, 'ctor, 'ty) t -> input -> validation_error
+val validate : ('meta, 'ctor, 'ty) t -> input -> error list
 
 (** [decode_and_validate schema input] returns the decoded and validated value
     of type ['ty] by decoding the [input] using the [schema] and running its
     validators.
-
-    The returned error is a {!type:validation_error}. If a field fails to
-    decode, the error of that field is the decode error. If a field decodes but
-    fails to validate, then the error is the validation error.
 
     Use [decode_and_validate] to combine the functions [decode] and [validate]
     and to either end up with the decoded value or all errors that happened
@@ -346,4 +341,4 @@ val validate : ('meta, 'ctor, 'ty) t -> input -> validation_error
 val decode_and_validate
   :  ('meta, 'ctor, 'ty) t
   -> input
-  -> ('ty, validation_error) Result.t
+  -> ('ty, error list) Result.t